requirements.txt
Django==1.8.2
djangorestframework>=3.1.1
nose==1.3.6
psycopg2==2.6
coverage==3.7.1
django-nose==1.4
django-admin-bootstrapped==2.5.0
